+++ mng/test.cpp
@@ -49,8 +49,8 @@ int main ()
 
 	block_type * block = new block_type;
 	STXXL_MSG(std::hex)
-	STXXL_MSG("Allocated block address    : "<<unsigned(block))
-	STXXL_MSG("Allocated block address + 1: "<<unsigned(block+1))
+	STXXL_MSG("Allocated block address    : "<<long(block))
+	STXXL_MSG("Allocated block address + 1: "<<long(block+1))
 	STXXL_MSG(std::dec)
 	unsigned i = 0;
 	for (i = 0; i < block_type::size; i++)
